requirements.txt
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dnspython==2.7.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
idna==3.10
ifaddr==0.2.0
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.47.1
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.3.1
pillow==10.4.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==0.12.0
Pympler==1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.28.2
rich==13.9.4
rpds-py==0.26.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.28.0
streamlit-webrtc==0.43.0
sympy==1.14.0
tenacity==8.5.0
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==4.3.1
ultralytics==8.0.201
urllib3==1.26.20
validators==0.35.0
zipp==3.23.0
